LayoutTests/http/tests/media/hls/hls-audio-tracks-language.html

<!DOCTYPE html>
<html>
    <head>
        <script src=../../media-resources/video-test.js></script>
        <script>
            if (window.testRunner) {
                testRunner.dumpAsText();
                testRunner.waitUntilDone();
            }

            function start() {
                video = document.getElementById('video');
                waitForEvent('canplaythrough', canplaythrough);
                video.src = "../resources/hls/audio-tracks.m3u8";
            }

            function canplaythrough() {
                testExpected("video.audioTracks.length", 3);
                testExpected("video.audioTracks[0].language", "en-US");
                testExpected("video.audioTracks[1].language", "fr-FR");
                testExpected("video.audioTracks[2].language", "es-US");
                testExpected("video.audioTracks[0].label", "English Sound");
                testExpected("video.audioTracks[1].label", "French Sound");
                testExpected("video.audioTracks[2].label", "Spanish Sound");
                endTest();
            }
        </script>
    </head>
    <body onload="start()">
        <video id="video"></video>
    </body>
</html>
